services/kafka_to_feature_store/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["j0n0m3 <44504165+j0n0m3@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
quixstreams = "^2.8.0"
loguru = "^0.7.2"
hopsworks = "^3.7.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
